services/multiplayer-phase-10-player-service/requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
flask_jwt_extended==4.6.0
requests==2.32.3
redis==5.0.8
python-dotenv==1.0.1
asyncio==3.4.3